ml_engine/requirements.txt
# ML Engine requirements
# Install with: pip install -r requirements.txt

# API server
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ML / NLP
transformers==4.40.0
torch==2.3.0
pyod==1.1.3
scikit-learn==1.4.2
numpy==1.26.4

# Kafka client (Redpanda-compatible)
kafka-python==2.0.2

# HTTP client (Grafana annotation pusher)
httpx==0.27.0

# Data validation (pulled in by fastapi, pinned for safety)
pydantic==2.7.1